Split V1/V2 into separate files and add -JsonSerializerOptions parameter for custom converter support
